lab5

+++ Lab5/CircleLawn.cpp
@@ -6,7 +6,7 @@ namespace lab5
 {
 	const double CircleLawn::PI = 3.14;
 
-	CircleLawn::CircleLawn(const unsigned int radius) :
+	CircleLawn::CircleLawn(const double radius) :
 		mRadius(radius)
 	{
 	}
@@ -21,6 +21,6 @@ namespace lab5
 	unsigned int CircleLawn::GetArea() const
 	{
 		const double area = mRadius * mRadius * PI;
-		return ceil(area);
+		return static_cast<unsigned int>(ceil(area));
 	}
 }
